+++ lib/Model/WhatsappMessageInteractiveActionSection.php
@@ -36,7 +36,7 @@ use \YCloud\Client\ObjectSerializer;
  * WhatsappMessageInteractiveActionSection Class Doc Comment
  *
  * @category Class
- * @description [WhatsApp Section Object](https://developers.facebook.com/docs/whatsapp/cloud-api/reference/messages#section-object).
+ * @description WhatsApp Message Interactive Section Object.
  * @package  YCloud\Client
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech