+++ lib/Model/WhatsappMessageTemplateComponentParameterActionSectionProductItem.php
@@ -217,7 +217,7 @@ class WhatsappMessageTemplateComponentParameterActionSectionProductItem implemen
     /**
      * Sets product_retailer_id
      *
-     * @param string|null $product_retailer_id SKU number of the item you want to appear in the section. SKU numbers are labeled as **Content ID** in the [Commerce Manager](https://business.facebook.com/commerce/).
+     * @param string|null $product_retailer_id SKU number of the item you want to appear in the section. SKU numbers are labeled as **Content ID** in the [Commerce Manager](https://business.facebook.com/commerce).
      *
      * @return self
      */